third_party/blink/web_tests/fast/files/blob-slice-test-expected.txt

Test Blob.slice().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Blob .slice(2, 3)
PASS blobContents is "2"
File .slice(2, 3)
PASS fileContents is "2"
Blob .slice(2, 12)
PASS blobContents is "23456789"
File .slice(2, 12)
PASS fileContents is "23456789"
Blob .slice(2, 2)
PASS blobContents is ""
File .slice(2, 2)
PASS fileContents is ""
Blob .slice(2, 1)
PASS blobContents is ""
File .slice(2, 1)
PASS fileContents is ""
Blob .slice(2, -12)
PASS blobContents is ""
File .slice(2, -12)
PASS fileContents is ""
Blob .slice(2, 2147483647)
PASS blobContents is "23456789"
File .slice(2, 2147483647)
PASS fileContents is "23456789"
Blob .slice(2, -2147483648)
PASS blobContents is ""
File .slice(2, -2147483648)
PASS fileContents is ""
Blob .slice(2, 9223372036854775000)
PASS blobContents is "23456789"
File .slice(2, 9223372036854775000)
PASS fileContents is "23456789"
Blob .slice(2, -9223372036854775000)
PASS blobContents is ""
File .slice(2, -9223372036854775000)
PASS fileContents is ""
Blob .slice(-2, -1)
PASS blobContents is "8"
File .slice(-2, -1)
PASS fileContents is "8"
Blob .slice(-2, -2)
PASS blobContents is ""
File .slice(-2, -2)
PASS fileContents is ""
Blob .slice(-2, -3)
PASS blobContents is ""
File .slice(-2, -3)
PASS fileContents is ""
Blob .slice(-2, -12)
PASS blobContents is ""
File .slice(-2, -12)
PASS fileContents is ""
Blob .slice(-2, 2147483647)
PASS blobContents is "89"
File .slice(-2, 2147483647)
PASS fileContents is "89"
Blob .slice(-2, -2147483648)
PASS blobContents is ""
File .slice(-2, -2147483648)
PASS fileContents is ""
Blob .slice(-2, 9223372036854775000)
PASS blobContents is "89"
File .slice(-2, 9223372036854775000)
PASS fileContents is "89"
Blob .slice(-2, -9223372036854775000)
PASS blobContents is ""
File .slice(-2, -9223372036854775000)
PASS fileContents is ""
Blob .slice(0)
PASS blobContents is "0123456789"
File .slice(0)
PASS fileContents is "0123456789"
Blob .slice(2)
PASS blobContents is "23456789"
File .slice(2)
PASS fileContents is "23456789"
Blob .slice(-2)
PASS blobContents is "89"
File .slice(-2)
PASS fileContents is "89"
Blob .slice(12)
PASS blobContents is ""
File .slice(12)
PASS fileContents is ""
Blob .slice(-12)
PASS blobContents is "0123456789"
File .slice(-12)
PASS fileContents is "0123456789"
Blob .slice(2147483647)
PASS blobContents is ""
File .slice(2147483647)
PASS fileContents is ""
Blob .slice(-2147483648)
PASS blobContents is "0123456789"
File .slice(-2147483648)
PASS fileContents is "0123456789"
Blob .slice(9223372036854775000)
PASS blobContents is ""
File .slice(9223372036854775000)
PASS fileContents is ""
Blob .slice(-9223372036854775000)
PASS blobContents is "0123456789"
File .slice(-9223372036854775000)
PASS fileContents is "0123456789"
Blob .slice()
PASS blobContents is "0123456789"
File .slice()
PASS fileContents is "0123456789"
PASS successfullyParsed is true

TEST COMPLETE
